Douwe's suggested edits and black to 120 per line
