+++ black.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 git ls-files \
   | egrep '\.py$' \
-  | xargs black --line-length 99 --skip-string-normalization $*
+  | xargs black --line-length 120 --skip-string-normalization $*
