+++ settings.yaml
@@ -11,11 +11,11 @@ start_grass_fraction: 0.3
 start_rock_fraction: 0.05
 start_water_fraction: 0.01
 grass_grow_per_turn: 3
-world_size: 35
-scale: 14
+world_size: 60
+scale: 10
 
 view_distance: 2
 steps_per_episode: 1000
 
-layers: [24, 16]
+layers: [24, 24]
