+++ simplegrid/deep_cow.py
@@ -1,7 +1,7 @@
 import os
 import numpy as np
 
-from simplegrid.cow import AbstractCreature, Action, MAX_ENERGY
+from simplegrid.abstractcreature import MAX_ENERGY, Action, AbstractCreature
 from simplegrid.dqn_agent import DQNAgent
 from simplegrid.map_feature import MapFeature
 
@@ -64,18 +64,14 @@ class DeepCow(AbstractCreature):
         self.prev_action_idx = self.action_idx
         self.state = self.to_internal_state(observation)
         if not DeepCow.agent:
-            DeepCow.agent = DQNAgent.from_dimensions(
-                len(self.state), layers=self.settings.layers, action_size=4
-            )
+            DeepCow.agent = DQNAgent.from_dimensions(len(self.state), layers=self.settings.layers, action_size=4)
         self.action_idx = DeepCow.agent.act(self.state)
         return Action(self.action_idx + 1)
 
     def learn(self, reward, done):
         self.reward = reward
         if self.prev_state is not None and self.state is not None:
-            DeepCow.agent.remember(
-                self.prev_state, self.prev_action_idx, self.prev_reward, self.state
-            )
+            DeepCow.agent.remember(self.prev_state, self.prev_action_idx, self.prev_reward, self.state)
             DeepCow.agent.replay()
         if done:
             DeepCow.agent.remember(self.state, self.action_idx, self.reward, None)
